Extract JSON credentials
